refactor(context): rename BindingKey.ts to binding-key.ts
